refactor: Point infrastructure 구성
