+++ src/main/java/com/example/surveyapp/domain/point/infrastructure/PointJpaRepository.java
@@ -1,14 +1,11 @@
-package com.example.surveyapp.domain.point.domain.repository;
+package com.example.surveyapp.domain.point.infrastructure;
 
 import com.example.surveyapp.domain.point.domain.model.entity.PointWallet;
-import jakarta.persistence.LockModeType;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Lock;
 
 import java.util.Optional;
 
-public interface PointRepository {
-//    @Lock(LockModeType.PESSIMISTIC_WRITE)
+public interface PointJpaRepository extends JpaRepository<PointWallet, Long> {
     Optional<PointWallet> findByUserId(Long userId);
     Boolean existsByUserId(Long userId);
