+++ src/main/java/com/example/surveyapp/domain/point/infrastructure/PointRepositoryImpl.java
@@ -1,0 +1,31 @@
+package com.example.surveyapp.domain.point.infrastructure;
+
+import com.example.surveyapp.domain.point.domain.model.entity.PointWallet;
+import com.example.surveyapp.domain.point.domain.repository.PointRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+import java.awt.*;
+import java.util.Optional;
+
+@Repository
+@RequiredArgsConstructor
+public class PointRepositoryImpl implements PointRepository {
+
+    private final PointJpaRepository pointJpaRepository;
+
+    @Override
+    public Optional<PointWallet> findByUserId(Long userId) {
+        return pointJpaRepository.findByUserId(userId);
+    }
+
+    @Override
+    public Boolean existsByUserId(Long userId) {
+        return pointJpaRepository.existsByUserId(userId);
+    }
+
+    @Override
+    public PointWallet save(PointWallet pointWallet) {
+        return pointJpaRepository.save(pointWallet);
+    }
+}
